requirements.txt
torch
transformers
pandas
llama_index
scikit-learn
numpy
matplotlib
sentence_transformers
langchain
openpyxl
javalang
chroma
llama_index-llms-anyscale
llama-index-vector-stores-chroma
llama-index-embeddings-huggingface
llama-index-llms-anyscale
openai
kuzu
llama-index-graph-stores-kuzu
llama-index-llms-cohere

